+++ ArcdpsLogManager/Controls/Filters/SquadCompositionFilterPanel.cs
@@ -74,6 +74,7 @@ namespace GW2Scratch.ArcdpsLogManager.Controls.Filters
 			BeginVertical(new Padding(5, 5), new Size(5, 5));
 			{
 				AddRow(null, ConstructResetButton(filterSnapshot), ConstructApplyButton(filters, filterSnapshot));
+				Add(null);
 			}
 			EndVertical();
 		}
@@ -105,20 +106,7 @@ namespace GW2Scratch.ArcdpsLogManager.Controls.Filters
 			};
 			
 			resetButton.Click += (_, _) => {
-				var filterGroups = new IReadOnlyList<PlayerCountFilter>[] {
-					filterSnapshot.CoreProfessionFilters,
-					filterSnapshot.HeartOfThornsSpecializationFilters,
-					filterSnapshot.PathOfFireSpecializationFilters,
-					filterSnapshot.EndOfDragonsSpecializationFilters,
-				};
-				foreach (var group in filterGroups)
-				{
-					foreach (var filter in group)
-					{
-						filter.PlayerCount = 0;
-						filter.FilterType = PlayerCountFilterType.GreaterOrEqual;
-					}
-				}
+				filterSnapshot.ResetToDefault();
 			};
 
 			filterSnapshot.PropertyChanged += (_, _) => {