+++ EVTCAnalytics/Model/Log.cs
@@ -5,6 +5,7 @@ using System.Threading;
 using GW2Scratch.EVTCAnalytics.Events;
 using GW2Scratch.EVTCAnalytics.GameData;
 using GW2Scratch.EVTCAnalytics.Model.Agents;
+using GW2Scratch.EVTCAnalytics.Model.Effects;
 using GW2Scratch.EVTCAnalytics.Model.Skills;
 using GW2Scratch.EVTCAnalytics.Processing;
 using GW2Scratch.EVTCAnalytics.Processing.Encounters;
@@ -30,6 +31,11 @@ namespace GW2Scratch.EVTCAnalytics.Model
 		/// Provides a list of all <see cref="Skill"/>s that appear in the log.
 		/// </summary>
 		public IReadOnlyList<Skill> Skills { get; }
+		
+		/// <summary>
+		/// Provides a list of all <see cref="Effect"/>s that appear in the log.
+		/// </summary>
+		public IReadOnlyList<Effect> Effects { get; }
 
 		/// <summary>
 		/// Provides a string with the version of arcdps used to record this log, prefixed with "EVTC".
@@ -112,11 +118,19 @@ namespace GW2Scratch.EVTCAnalytics.Model
 		/// Added in arcdps 20211214.
 		/// </remarks>
 		public InstanceStart InstanceStart { get; }
+		
+		/// <summary>
+		/// Provides a list of all errors logged by arcdps.
+		/// </summary>
+		/// <remarks>
+		/// Added in arcdps 20200513.
+		/// </remarks>
+		public IReadOnlyList<LogError> Errors { get; }
 
 		/// <summary>
 		/// Creates a new instance of a <see cref="Log"/>.
 		/// </summary>
-		/// <param name="state">The context of a <see cref="LogProcessor"/>.</param>
+		/// <param name="state">The state of a <see cref="LogProcessor"/> which is consumed.</param>
 		internal Log(LogProcessorState state)
 		{
 			MainTarget = state.MainTarget;
@@ -135,6 +149,8 @@ namespace GW2Scratch.EVTCAnalytics.Model
 			Events = state.Events;
 			Agents = state.Agents;
 			Skills = state.Skills;
+			Effects = state.EffectsById.Values.OrderBy(x => x.Id).ToList();
+			Errors = state.Errors;
 			InstanceStart = state.InstanceStart;
 		}
 
@@ -142,9 +158,9 @@ namespace GW2Scratch.EVTCAnalytics.Model
 		/// Creates a new instance of a <see cref="Log"/> without requiring a <see cref="LogProcessorState"/>.
 		/// </summary>
 		internal Log(Agent mainTarget, LogType logType, IEnumerable<Event> events, IEnumerable<Agent> agents,
-			IEnumerable<Skill> skills, IEncounterData encounterData,
-			GameLanguage gameLanguage, string evtcVersion, LogTime startTime, LogTime endTime,
-			Player pointOfView, int? language, int? gameBuild, int? gameShardId, int? mapId,
+			IEnumerable<Skill> skills, IEnumerable<Effect> effects, IEnumerable<LogError> errors,
+			IEncounterData encounterData, GameLanguage gameLanguage, string evtcVersion, LogTime startTime,
+			LogTime endTime, Player pointOfView, int? language, int? gameBuild, int? gameShardId, int? mapId,
 			InstanceStart instanceStart)
 		{
 			MainTarget = mainTarget;
@@ -163,6 +179,8 @@ namespace GW2Scratch.EVTCAnalytics.Model
 			Events = events as Event[] ?? events.ToArray();
 			Agents = agents as Agent[] ?? agents.ToArray();
 			Skills = skills as Skill[] ?? skills.ToArray();
+			Effects = effects as Effect[] ?? effects.ToArray();
+			Errors = errors as LogError[] ?? errors.ToArray();
 		}
 	}
 }