+++ EVTCAnalytics/Parsed/Enums/Activation.cs
@@ -5,13 +5,11 @@
 	/// </summary>
 	public enum Activation : byte
 	{
-		None         = 0,
-		Normal       = 1,
-		Quickness    = 2,
-		CancelFire   = 3,
+		None = 0,
+		Normal = 1,
+		Quickness = 2,
+		CancelFire = 3,
 		CancelCancel = 4,
-		Reset        = 5,
-
-		Unknown
+		Reset = 5,
 	}
 }